actualizacion logica pedidos y productos por pedido

+++ backend/PedidosSYAC.Common/Dto/Productos/ProductoDto.cs
@@ -15,7 +15,7 @@ namespace PedidosSYAC.Common.Dto.Productos
         [MaxLength(50)]
         public string Nombre { get; set; }
         public int Cantidad { get; set; }
-        public int ValorUnitario { get; set; }
+        public decimal ValorUnitario { get; set; }
 
     }
 }
